modules/fantasy-football/functions/fpl_weekly_update/requirements.txt
requests>=2.31.0
cognite-sdk>=7.0.0
numpy>=1.24.0

